Add admin promotion duplication functionality
Adds ‘duplicate’ button to promotion page in admin. Uses a get route to build the promotion on the page, and a POST request to create to create the promotion. Promotion rules and actions are not able to be edited in the duplicate page, but can be edited like regular promotions after create

* Adds get route for duplicate promo
* Adds link in admin_promotions_path to duplicate page
* Adds duplicate page and styles
* Remove unused conditionals promotion action partials